requirements-dev.txt
pytest==8.4.2
moto[s3]>=5.0.0
ruff>=0.4.0
httpx>=0.24.0,<0.28.0
boto3-stubs[dynamodb,s3]
pyright
pytest-cov
email-validator>=2.0.0
PyJWT>=2.8.0
